src/llms/requirements.txt
transformers==4.46.3
torch==2.5.1+cu121
numpy==1.26.4
datasets==3.2.0
google-generativeai>=0.7.2
